Cgarciag/pytest setup (#18)
