--- pyproject.toml
@@ -1,7 +1,0 @@
-[tool.black]
-line-length = 79
-target-version = ['py310']
-
-[tool.isort]
-profile = "black"
-line_length = 79
