+++ setup.cfg
@@ -1,0 +1,239 @@
+[metadata]
+name = maternalhealth
+description = slap that like button
+author = Carlos Garcia
+license = MIT
+license_file = LICENSE
+platforms = unix, linux, osx, cygwin, win32
+classifiers =
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+
+[options]
+package_dir=
+    =src
+packages=find:
+install_requires =
+    aiohappyeyeballs==2.4.3
+    aiohttp==3.10.9
+    aiohttp-retry==2.8.3
+    aiosignal==1.3.1
+    alembic==1.13.3
+    amqp==5.2.0
+    annotated-types==0.7.0
+    antlr4-python3-runtime==4.9.3
+    appdirs==1.4.4
+    appnope==0.1.4
+    asttokens==2.4.1
+    async-timeout==4.0.3
+    asyncssh==2.17.0
+    atpublic==5.0
+    attrs==24.2.0
+    billiard==4.2.1
+    blinker==1.8.2
+    cachetools==5.5.0
+    celery==5.4.0
+    certifi==2024.8.30
+    cffi==1.17.1
+    cfgv==3.4.0
+    charset-normalizer==3.3.2
+    click==8.1.7
+    click-didyoumean==0.3.1
+    click-plugins==1.1.1
+    click-repl==0.3.0
+    cloudpickle==3.1.0
+    colorama==0.4.6
+    comm==0.2.2
+    configobj==5.0.9
+    contourpy==1.3.0
+    cryptography==43.0.1
+    cycler==0.12.1
+    databricks-sdk==0.35.0
+    debugpy==1.8.6
+    decorator==5.1.1
+    Deprecated==1.2.14
+    dictdiffer==0.9.0
+    diskcache==5.6.3
+    distlib==0.3.9
+    distro==1.9.0
+    docker==7.1.0
+    dpath==2.2.0
+    dulwich==0.22.1
+    dvc==3.55.2
+    dvc-data==3.16.6
+    dvc-gdrive==3.0.1
+    dvc-http==2.32.0
+    dvc-objects==5.1.0
+    dvc-render==1.0.2
+    dvc-studio-client==0.21.0
+    dvc-task==0.40.1
+    entrypoints==0.4
+    exceptiongroup==1.2.2
+    executing==2.1.0
+    filelock==3.16.1
+    Flask==3.0.3
+    flatten-dict==0.4.2
+    flufl.lock==8.1.0
+    fonttools==4.54.1
+    frozenlist==1.4.1
+    fsspec==2024.9.0
+    funcy==2.0
+    gitdb==4.0.11
+    GitPython==3.1.43
+    google-api-core==2.21.0
+    google-api-python-client==2.149.0
+    google-auth==2.35.0
+    google-auth-httplib2==0.2.0
+    googleapis-common-protos==1.65.0
+    grandalf==0.8
+    graphene==3.4
+    graphql-core==3.2.5
+    graphql-relay==3.2.0
+    greenlet==3.1.1
+    gto==1.7.1
+    gunicorn==23.0.0
+    httplib2==0.22.0
+    hydra-core==1.3.2
+    identify==2.6.1
+    idna==3.10
+    importlib_metadata==8.4.0
+    iniconfig==2.0.0
+    ipykernel==6.29.5
+    ipython==8.27.0
+    iterative-telemetry==0.0.9
+    itsdangerous==2.2.0
+    jedi==0.19.1
+    Jinja2==3.1.4
+    joblib==1.4.2
+    jupyter_client==8.6.3
+    jupyter_core==5.7.2
+    kiwisolver==1.4.7
+    kombu==5.4.2
+    Mako==1.3.5
+    Markdown==3.7
+    markdown-it-py==3.0.0
+    MarkupSafe==3.0.2
+    -e git+https://github.com/Immutare/mlops-project-team13.git@e7dcd3abc97168ce2d7756a6e651a8171cbb1d55#egg=maternalhealth
+    matplotlib==3.9.2
+    matplotlib-inline==0.1.7
+    mdurl==0.1.2
+    mlflow==2.17.0
+    mlflow-skinny==2.17.0
+    multidict==6.1.0
+    nest-asyncio==1.6.0
+    networkx==3.3
+    nodeenv==1.9.1
+    numpy==2.1.1
+    oauth2client==4.1.3
+    omegaconf==2.3.0
+    opentelemetry-api==1.27.0
+    opentelemetry-sdk==1.27.0
+    opentelemetry-semantic-conventions==0.48b0
+    orjson==3.10.7
+    packaging==24.1
+    pandas==2.2.3
+    parso==0.8.4
+    pathspec==0.12.1
+    pexpect==4.9.0
+    pillow==10.4.0
+    platformdirs==3.11.0
+    pluggy==1.5.0
+    pre_commit==4.0.1
+    prompt_toolkit==3.0.48
+    proto-plus==1.24.0
+    protobuf==5.28.2
+    psutil==6.0.0
+    ptyprocess==0.7.0
+    pure_eval==0.2.3
+    pyarrow==17.0.0
+    pyasn1==0.6.1
+    pyasn1_modules==0.4.1
+    pycparser==2.22
+    pydantic==2.9.2
+    pydantic_core==2.23.4
+    pydot==3.0.2
+    PyDrive2==1.20.0
+    pygit2==1.15.1
+    Pygments==2.18.0
+    pygtrie==2.5.0
+    pyOpenSSL==24.2.1
+    pyparsing==3.1.4
+    pytest==8.3.3
+    python-dateutil==2.9.0.post0
+    pytz==2024.2
+    pywin32==306
+    PyYAML==6.0.2
+    pyzmq==26.2.0
+    requests==2.32.3
+    rich==13.9.2
+    rsa==4.9
+    ruamel.yaml==0.18.6
+    ruamel.yaml.clib==0.2.8
+    scikit-learn==1.5.2
+    scipy==1.14.1
+    scmrepo==3.3.8
+    seaborn==0.13.2
+    semver==3.0.2
+    setuptools==75.1.0
+    shellingham==1.5.4
+    shortuuid==1.0.13
+    shtab==1.7.1
+    six==1.16.0
+    smmap==5.0.1
+    SQLAlchemy==2.0.36
+    sqlparse==0.5.1
+    sqltrie==0.11.1
+    stack-data==0.6.3
+    tabulate==0.9.0
+    threadpoolctl==3.5.0
+    tomli==2.0.2
+    tomlkit==0.13.2
+    tornado==6.4.1
+    tqdm==4.66.5
+    traitlets==5.14.3
+    typer==0.12.5
+    typing_extensions==4.12.2
+    tzdata==2024.2
+    ucimlrepo==0.0.7
+    uritemplate==4.1.1
+    urllib3==2.2.3
+    vine==5.1.0
+    virtualenv==20.27.1
+    voluptuous==0.15.2
+    waitress==3.0.0
+    wcwidth==0.2.13
+    Werkzeug==3.0.4
+    wrapt==1.16.0
+    xgboost==2.1.1
+    yarl==1.13.1
+    zc.lockfile==3.0.post1
+    zipp==3.20.2
+python_requires = >=3.6
+zip_safe = no
+
+[options.extras_require]
+testing =
+    pytest==8.3.3
+    pytest-cov>=2.0
+    mypy>=0.910
+    flake8>=3.9
+    tox>=3.24
+
+[options.package_data]
+maternalhealth = py.typed
+
+
+[options.packages.find]
+where=src
+
+[flake8]
+max-line-length = 79
+
+[tool:pytest]
+addopts = --ignore=tests/test_modelclassifier.py
+testpaths =
+    tests
