+++ src/maternalhealth.egg-info/PKG-INFO
@@ -1,0 +1,204 @@
+Metadata-Version: 2.1
+Name: maternalhealth
+Version: 0.0.0
+Summary: slap that like button
+Author: Carlos Garcia
+License: MIT
+Platform: unix
+Platform: linux
+Platform: osx
+Platform: cygwin
+Platform: win32
+Classifier: Programming Language :: Python :: 3
+Classifier: Programming Language :: Python :: 3 :: Only
+Classifier: Programming Language :: Python :: 3.6
+Classifier: Programming Language :: Python :: 3.7
+Classifier: Programming Language :: Python :: 3.8
+Classifier: Programming Language :: Python :: 3.9
+Requires-Python: >=3.6
+Requires-Dist: aiohappyeyeballs==2.4.3
+Requires-Dist: aiohttp==3.10.9
+Requires-Dist: aiohttp-retry==2.8.3
+Requires-Dist: aiosignal==1.3.1
+Requires-Dist: alembic==1.13.3
+Requires-Dist: amqp==5.2.0
+Requires-Dist: annotated-types==0.7.0
+Requires-Dist: antlr4-python3-runtime==4.9.3
+Requires-Dist: appdirs==1.4.4
+Requires-Dist: asttokens==2.4.1
+Requires-Dist: asyncssh==2.17.0
+Requires-Dist: atpublic==5.0
+Requires-Dist: attrs==24.2.0
+Requires-Dist: billiard==4.2.1
+Requires-Dist: blinker==1.8.2
+Requires-Dist: cachetools==5.5.0
+Requires-Dist: celery==5.4.0
+Requires-Dist: certifi==2024.8.30
+Requires-Dist: cffi==1.17.1
+Requires-Dist: charset-normalizer==3.3.2
+Requires-Dist: click==8.1.7
+Requires-Dist: click-didyoumean==0.3.1
+Requires-Dist: click-plugins==1.1.1
+Requires-Dist: click-repl==0.3.0
+Requires-Dist: cloudpickle==3.1.0
+Requires-Dist: colorama==0.4.6
+Requires-Dist: comm==0.2.2
+Requires-Dist: configobj==5.0.9
+Requires-Dist: contourpy==1.3.0
+Requires-Dist: cryptography==43.0.1
+Requires-Dist: cycler==0.12.1
+Requires-Dist: databricks-sdk==0.35.0
+Requires-Dist: debugpy==1.8.6
+Requires-Dist: decorator==5.1.1
+Requires-Dist: Deprecated==1.2.14
+Requires-Dist: dictdiffer==0.9.0
+Requires-Dist: diskcache==5.6.3
+Requires-Dist: distro==1.9.0
+Requires-Dist: docker==7.1.0
+Requires-Dist: dpath==2.2.0
+Requires-Dist: dulwich==0.22.1
+Requires-Dist: dvc==3.55.2
+Requires-Dist: dvc-data==3.16.6
+Requires-Dist: dvc-gdrive==3.0.1
+Requires-Dist: dvc-http==2.32.0
+Requires-Dist: dvc-objects==5.1.0
+Requires-Dist: dvc-render==1.0.2
+Requires-Dist: dvc-studio-client==0.21.0
+Requires-Dist: dvc-task==0.40.1
+Requires-Dist: entrypoints==0.4
+Requires-Dist: executing==2.1.0
+Requires-Dist: filelock==3.16.1
+Requires-Dist: Flask==3.0.3
+Requires-Dist: flatten-dict==0.4.2
+Requires-Dist: flufl.lock==8.1.0
+Requires-Dist: fonttools==4.54.1
+Requires-Dist: frozenlist==1.4.1
+Requires-Dist: fsspec==2024.9.0
+Requires-Dist: funcy==2.0
+Requires-Dist: gitdb==4.0.11
+Requires-Dist: GitPython==3.1.43
+Requires-Dist: google-api-core==2.21.0
+Requires-Dist: google-api-python-client==2.149.0
+Requires-Dist: google-auth==2.35.0
+Requires-Dist: google-auth-httplib2==0.2.0
+Requires-Dist: googleapis-common-protos==1.65.0
+Requires-Dist: grandalf==0.8
+Requires-Dist: graphene==3.4
+Requires-Dist: graphql-core==3.2.5
+Requires-Dist: graphql-relay==3.2.0
+Requires-Dist: greenlet==3.1.1
+Requires-Dist: gto==1.7.1
+Requires-Dist: httplib2==0.22.0
+Requires-Dist: hydra-core==1.3.2
+Requires-Dist: idna==3.10
+Requires-Dist: importlib_metadata==8.4.0
+Requires-Dist: iniconfig==2.0.0
+Requires-Dist: ipykernel==6.29.5
+Requires-Dist: ipython==8.27.0
+Requires-Dist: iterative-telemetry==0.0.9
+Requires-Dist: itsdangerous==2.2.0
+Requires-Dist: jedi==0.19.1
+Requires-Dist: Jinja2==3.1.4
+Requires-Dist: joblib==1.4.2
+Requires-Dist: jupyter_client==8.6.3
+Requires-Dist: jupyter_core==5.7.2
+Requires-Dist: kiwisolver==1.4.7
+Requires-Dist: kombu==5.4.2
+Requires-Dist: Mako==1.3.5
+Requires-Dist: Markdown==3.7
+Requires-Dist: markdown-it-py==3.0.0
+Requires-Dist: MarkupSafe==3.0.2
+Requires-Dist: matplotlib==3.9.2
+Requires-Dist: matplotlib-inline==0.1.7
+Requires-Dist: mdurl==0.1.2
+Requires-Dist: mlflow==2.17.0
+Requires-Dist: mlflow-skinny==2.17.0
+Requires-Dist: multidict==6.1.0
+Requires-Dist: nest-asyncio==1.6.0
+Requires-Dist: networkx==3.3
+Requires-Dist: numpy==2.1.1
+Requires-Dist: oauth2client==4.1.3
+Requires-Dist: omegaconf==2.3.0
+Requires-Dist: opentelemetry-api==1.27.0
+Requires-Dist: opentelemetry-sdk==1.27.0
+Requires-Dist: opentelemetry-semantic-conventions==0.48b0
+Requires-Dist: orjson==3.10.7
+Requires-Dist: packaging==24.1
+Requires-Dist: pandas==2.2.3
+Requires-Dist: parso==0.8.4
+Requires-Dist: pathspec==0.12.1
+Requires-Dist: pillow==10.4.0
+Requires-Dist: platformdirs==3.11.0
+Requires-Dist: pluggy==1.5.0
+Requires-Dist: prompt_toolkit==3.0.48
+Requires-Dist: proto-plus==1.24.0
+Requires-Dist: protobuf==5.28.2
+Requires-Dist: psutil==6.0.0
+Requires-Dist: pure_eval==0.2.3
+Requires-Dist: pyarrow==17.0.0
+Requires-Dist: pyasn1==0.6.1
+Requires-Dist: pyasn1_modules==0.4.1
+Requires-Dist: pycparser==2.22
+Requires-Dist: pydantic==2.9.2
+Requires-Dist: pydantic_core==2.23.4
+Requires-Dist: pydot==3.0.2
+Requires-Dist: PyDrive2==1.20.0
+Requires-Dist: pygit2==1.15.1
+Requires-Dist: Pygments==2.18.0
+Requires-Dist: pygtrie==2.5.0
+Requires-Dist: pyOpenSSL==24.2.1
+Requires-Dist: pyparsing==3.1.4
+Requires-Dist: pytest==8.3.3
+Requires-Dist: python-dateutil==2.9.0.post0
+Requires-Dist: pytz==2024.2
+Requires-Dist: pywin32==306; platform_system == "Windows"
+Requires-Dist: PyYAML==6.0.2
+Requires-Dist: pyzmq==26.2.0
+Requires-Dist: requests==2.32.3
+Requires-Dist: rich==13.9.2
+Requires-Dist: rsa==4.9
+Requires-Dist: ruamel.yaml==0.18.6
+Requires-Dist: ruamel.yaml.clib==0.2.8
+Requires-Dist: scikit-learn==1.5.2
+Requires-Dist: scipy==1.14.1
+Requires-Dist: scmrepo==3.3.8
+Requires-Dist: seaborn==0.13.2
+Requires-Dist: semver==3.0.2
+Requires-Dist: setuptools==75.1.0
+Requires-Dist: shellingham==1.5.4
+Requires-Dist: shortuuid==1.0.13
+Requires-Dist: shtab==1.7.1
+Requires-Dist: six==1.16.0
+Requires-Dist: smmap==5.0.1
+Requires-Dist: SQLAlchemy==2.0.36
+Requires-Dist: sqlparse==0.5.1
+Requires-Dist: sqltrie==0.11.1
+Requires-Dist: stack-data==0.6.3
+Requires-Dist: tabulate==0.9.0
+Requires-Dist: threadpoolctl==3.5.0
+Requires-Dist: tomlkit==0.13.2
+Requires-Dist: tornado==6.4.1
+Requires-Dist: tqdm==4.66.5
+Requires-Dist: traitlets==5.14.3
+Requires-Dist: typer==0.12.5
+Requires-Dist: typing_extensions==4.12.2
+Requires-Dist: tzdata==2024.2
+Requires-Dist: ucimlrepo==0.0.7
+Requires-Dist: uritemplate==4.1.1
+Requires-Dist: urllib3==2.2.3
+Requires-Dist: vine==5.1.0
+Requires-Dist: voluptuous==0.15.2
+Requires-Dist: waitress==3.0.0
+Requires-Dist: wcwidth==0.2.13
+Requires-Dist: Werkzeug==3.0.4
+Requires-Dist: wrapt==1.16.0
+Requires-Dist: yarl==1.13.1
+Requires-Dist: zc.lockfile==3.0.post1
+Requires-Dist: zipp==3.20.2
+Requires-Dist: xgboost==2.1.1
+Provides-Extra: testing
+Requires-Dist: pytest==8.3.3; extra == "testing"
+Requires-Dist: pytest-cov>=2.0; extra == "testing"
+Requires-Dist: mypy>=0.910; extra == "testing"
+Requires-Dist: flake8>=3.9; extra == "testing"
+Requires-Dist: tox>=3.24; extra == "testing"
