+++ src/maternalhealth.egg-info/SOURCES.txt
@@ -1,0 +1,24 @@
+setup.cfg
+setup.py
+src/maternalhealth/__init__.py
+src/maternalhealth.egg-info/PKG-INFO
+src/maternalhealth.egg-info/SOURCES.txt
+src/maternalhealth.egg-info/dependency_links.txt
+src/maternalhealth.egg-info/not-zip-safe
+src/maternalhealth.egg-info/requires.txt
+src/maternalhealth.egg-info/top_level.txt
+src/maternalhealth/pipeline/__init__.py
+src/maternalhealth/pipeline/evaluate.py
+src/maternalhealth/pipeline/prepare.py
+src/maternalhealth/pipeline/preprocess.py
+src/maternalhealth/pipeline/model/__init__.py
+src/maternalhealth/pipeline/model/classifiermodel.py
+src/maternalhealth/pipeline/model/informalmodelinterface.py
+src/maternalhealth/pipeline/training/__init__.py
+src/maternalhealth/pipeline/training/train_decisiontree.py
+src/maternalhealth/pipeline/training/train_knn.py
+src/maternalhealth/pipeline/training/train_logistic.py
+src/maternalhealth/pipeline/training/train_random_forest.py
+src/maternalhealth/pipeline/training/train_svm.py
+src/maternalhealth/pipeline/training/train_xgboost.py
+tests/test_modelclassifier.py
