+++ src/maternalhealth.egg-info/requires.txt
@@ -7,9 +7,7 @@ amqp==5.2.0
 annotated-types==0.7.0
 antlr4-python3-runtime==4.9.3
 appdirs==1.4.4
-appnope==0.1.4
 asttokens==2.4.1
-async-timeout==4.0.3
 asyncssh==2.17.0
 atpublic==5.0
 attrs==24.2.0
@@ -19,7 +17,6 @@ cachetools==5.5.0
 celery==5.4.0
 certifi==2024.8.30
 cffi==1.17.1
-cfgv==3.4.0
 charset-normalizer==3.3.2
 click==8.1.7
 click-didyoumean==0.3.1
@@ -38,7 +35,6 @@ decorator==5.1.1
 Deprecated==1.2.14
 dictdiffer==0.9.0
 diskcache==5.6.3
-distlib==0.3.9
 distro==1.9.0
 docker==7.1.0
 dpath==2.2.0
@@ -52,7 +48,6 @@ dvc-render==1.0.2
 dvc-studio-client==0.21.0
 dvc-task==0.40.1
 entrypoints==0.4
-exceptiongroup==1.2.2
 executing==2.1.0
 filelock==3.16.1
 Flask==3.0.3
@@ -75,10 +70,8 @@ graphql-core==3.2.5
 graphql-relay==3.2.0
 greenlet==3.1.1
 gto==1.7.1
-gunicorn==23.0.0
 httplib2==0.22.0
 hydra-core==1.3.2
-identify==2.6.1
 idna==3.10
 importlib_metadata==8.4.0
 iniconfig==2.0.0
@@ -97,7 +90,6 @@ Mako==1.3.5
 Markdown==3.7
 markdown-it-py==3.0.0
 MarkupSafe==3.0.2
--e git+https://github.com/Immutare/mlops-project-team13.git@e7dcd3abc97168ce2d7756a6e651a8171cbb1d55#egg=maternalhealth
 matplotlib==3.9.2
 matplotlib-inline==0.1.7
 mdurl==0.1.2
@@ -106,7 +98,6 @@ mlflow-skinny==2.17.0
 multidict==6.1.0
 nest-asyncio==1.6.0
 networkx==3.3
-nodeenv==1.9.1
 numpy==2.1.1
 oauth2client==4.1.3
 omegaconf==2.3.0
@@ -118,16 +109,13 @@ packaging==24.1
 pandas==2.2.3
 parso==0.8.4
 pathspec==0.12.1
-pexpect==4.9.0
 pillow==10.4.0
 platformdirs==3.11.0
 pluggy==1.5.0
-pre_commit==4.0.1
 prompt_toolkit==3.0.48
 proto-plus==1.24.0
 protobuf==5.28.2
 psutil==6.0.0
-ptyprocess==0.7.0
 pure_eval==0.2.3
 pyarrow==17.0.0
 pyasn1==0.6.1
@@ -145,7 +133,6 @@ pyparsing==3.1.4
 pytest==8.3.3
 python-dateutil==2.9.0.post0
 pytz==2024.2
-pywin32==306; platform_system == "Windows" 
 PyYAML==6.0.2
 pyzmq==26.2.0
 requests==2.32.3
@@ -170,7 +157,6 @@ sqltrie==0.11.1
 stack-data==0.6.3
 tabulate==0.9.0
 threadpoolctl==3.5.0
-tomli==2.0.2
 tomlkit==0.13.2
 tornado==6.4.1
 tqdm==4.66.5
@@ -182,13 +168,22 @@ ucimlrepo==0.0.7
 uritemplate==4.1.1
 urllib3==2.2.3
 vine==5.1.0
-virtualenv==20.27.1
 voluptuous==0.15.2
 waitress==3.0.0
 wcwidth==0.2.13
 Werkzeug==3.0.4
 wrapt==1.16.0
-xgboost==2.1.1
 yarl==1.13.1
 zc.lockfile==3.0.post1
 zipp==3.20.2
+xgboost==2.1.1
+
+[:platform_system == "Windows"]
+pywin32==306
+
+[testing]
+pytest==8.3.3
+pytest-cov>=2.0
+mypy>=0.910
+flake8>=3.9
+tox>=3.24
