+++ src/maternalhealth/pipeline/training/train_random_forest.py
@@ -5,14 +5,15 @@ import joblib
 import mlflow
 import pandas as pd
 import yaml
+from sklearn.ensemble import RandomForestClassifier
 from sklearn.model_selection import GridSearchCV
-from sklearn.tree import DecisionTreeClassifier
 
 
-def trainTree(X, Y, **kwargs):
+def train_random_forest(X, Y, **kwargs):
     """
-    Función para entrenar un modelo DecisionTreeClassifier utilizando búsqueda
-    de hiperparámetros con GridSearchCV.
+    Función para entrenar un modelo RandomForestClassifier utilizando búsqueda
+    de hiperparámetros
+    con GridSearchCV.
 
     Parámetros:
     -----------
@@ -22,28 +23,34 @@ def trainTree(X, Y, **kwargs):
     Y : numpy array o pandas Series
         Etiquetas (labels) correspondientes a los datos de entrenamiento.
 
+    n_estimators : list
+        Número de árboles en el bosque.
     criterion : list
         Funcion para medir la calidad de un split.
     max_depth : list
         Profundidad maxima de un arbol
-    min_samples_split : int
-        Minimo de muestras requeridas para dividir un nodo
+    max_features : list
+        Número de características a considerar en cada split.
+    min_samples_leaf : list
+        Minimo de muestras requeridas para ser una hoja.
     Retorno:
     --------
-    best_decision_tree_model : DecisionTreeClassifier
-        El mejor modelo DecisionTreeClassifier encontrado durante la búsqueda
+    best_random_forest_model : RandomForestClassifier
+        El mejor modelo RandomForestClassifier encontrado durante la búsqueda
         de hiperparámetros utilizando GridSearchCV.
     """
 
     parameters = {
+        "n_estimators": kwargs["n_estimators"],
         "criterion": kwargs["criterion"],
         "max_depth": kwargs["max_depth"],
-        "min_samples_split": kwargs["min_samples_split"],
+        "max_features": kwargs["max_features"],
+        "min_samples_leaf": kwargs["min_samples_leaf"],
     }
 
     # Realizar búsqueda de hiperparámetros usando GridSearchCV
     grid_search = GridSearchCV(
-        estimator=DecisionTreeClassifier(random_state=kwargs["seed"]),
+        estimator=RandomForestClassifier(random_state=kwargs["seed"]),
         param_grid=parameters,
         scoring="f1_weighted",
         verbose=0,
@@ -54,18 +61,16 @@ def trainTree(X, Y, **kwargs):
     grid_search.fit(X, Y)
 
     # Obtener los mejores hiperparámetros encontrados
-    best_decision_tree_model = grid_search.best_estimator_
+    best_random_forest_model = grid_search.best_estimator_
     best_params = grid_search.best_params_
 
-    return best_decision_tree_model, best_params
+    return best_random_forest_model, best_params
 
 
 def main():
     if len(sys.argv) != 3:
         sys.stderr.write("Arguments error. Usage:\n")
-        sys.stderr.write(
-            "\tpython train.py preprocessed-dir-path model-dir-path\n"
-        )
+        sys.stderr.write("\tpython train.py preprocessed-dir-path model-dir-path\n")
         sys.exit(1)
 
     in_path = sys.argv[1]  # Ruta de los datos preprocesados
@@ -83,55 +88,42 @@ def main():
     y_train = train_data["0"]
 
     # Cargar los parámetros del archivo params.yaml
-    params = yaml.safe_load(open("params.yaml"))["train_decision_tree"]
+    params = yaml.safe_load(open("params.yaml"))["train_random_forest"]
     # version_label = yaml.safe_load(open("params.yaml"))["version_label"]
     mlflow_params = yaml.safe_load(open("params.yaml"))["mlflow"]
-
-    # Inicializar la clase ClassifierModel con la versión y semilla
-    # classifier_model = trainTree(X_train, y_train, **params)
 
     # Inicializar MLFlow
     mlflow.set_experiment(mlflow_params["experiment_name"])
     mlflow.set_tracking_uri(mlflow_params["tracking_uri"])
 
     # Asignar el nombre del run usando el formato definido en params.yaml
-    run_name = mlflow_params["run_name"].format(model="Decision Tree")
+    run_name = mlflow_params["run_name"].format(model="Random Forest")
 
-    # Entrenar y guardar el modelo DecisionTree usando los parámetros de
+    # Entrenar y guardar el modelo RandomForest usando los parámetros de
     # `params.yaml`
     with mlflow.start_run(run_name=run_name) as run:
-        decision_tree_model, best_params = trainTree(
+        random_forest_model, best_params = train_random_forest(
             X_train, y_train, **params
         )
 
         # Guardar los mejores hiperparámetros en MLflow
         mlflow.log_params(best_params)
 
-        # Verificar si el modelo tiene un atributo best_params_
-        # if hasattr(decision_tree_model, "best_params_"):
-        # Guardar solo los mejores hiperparámetros en MLflow
-        #    mlflow.log_params(decision_tree_model.best_params_)
-        # else:
-        # Si no hay una búsqueda de hiperparámetros, guardar manualmente
-        #    mlflow.log_param("n_neighbors", params["n_neighbors"])
-        #    mlflow.log_param("weights", params["weights"])
-        #    mlflow.log_param("distance_metric", params["p"])
-
         # Guardar el modelo en MLFlow
-        mlflow.sklearn.log_model(decision_tree_model, "DT_model")
+        mlflow.sklearn.log_model(random_forest_model, "RF_model")
 
         # Guardar el run_id para compartirlo con evaluate.py
         run_id = run.info.run_id
         with open(
-            os.path.join(out_path, "mlflow_run_id_decision_tree_model.txt"),
+            os.path.join(out_path, "mlflow_run_id_random_forest_model.txt"),
             "w",
         ) as f:
             f.write(run_id)
 
     # Guardar el modelo en el directorio de salida
-    model_output_path = os.path.join(out_path, "decision_tree_model.pkl")
-    joblib.dump(decision_tree_model, model_output_path)
-    sys.stderr.write(f"Modelo Decision Tree guardado en {model_output_path}\n")
+    model_output_path = os.path.join(out_path, "random_forest_model.pkl")
+    joblib.dump(random_forest_model, model_output_path)
+    sys.stdout.write(f"Modelo Random Forest guardado en {model_output_path}\n")
 
 
 if __name__ == "__main__":
