+++ src/maternalhealth/pipeline/training/train_svm.py
@@ -5,14 +5,13 @@ import joblib
 import mlflow
 import pandas as pd
 import yaml
-from sklearn.ensemble import RandomForestClassifier
 from sklearn.model_selection import GridSearchCV
+from sklearn.svm import SVC
 
 
-def train_random_forest(X, Y, **kwargs):
+def trainSVM(X, Y, **kwargs):
     """
-    Función para entrenar un modelo RandomForestClassifier utilizando búsqueda
-    de hiperparámetros
+    Función para entrenar un modelo SVM utilizando búsqueda de hiperparámetros
     con GridSearchCV.
 
     Parámetros:
@@ -23,34 +22,29 @@ def train_random_forest(X, Y, **kwargs):
     Y : numpy array o pandas Series
         Etiquetas (labels) correspondientes a los datos de entrenamiento.
 
-    n_estimators : list
-        Número de árboles en el bosque.
-    criterion : list
-        Funcion para medir la calidad de un split.
-    max_depth : list
-        Profundidad maxima de un arbol
-    max_features : list
-        Número de características a considerar en cada split.
-    min_samples_leaf : list
-        Minimo de muestras requeridas para ser una hoja.
+    C : list
+        Valores de regularización del modelo SVM.
+    kernel : list
+        Tipo de kernel a usar (e.g., 'linear', 'rbf', etc.)
+    gamma : list
+        Parámetro de kernel para modelos no lineales.
+
     Retorno:
     --------
-    best_random_forest_model : RandomForestClassifier
-        El mejor modelo RandomForestClassifier encontrado durante la búsqueda
-        de hiperparámetros utilizando GridSearchCV.
+    best_svm_model : SVC
+        El mejor modelo SVM encontrado durante la búsqueda de hiperparámetros
+        utilizando GridSearchCV.
     """
 
     parameters = {
-        "n_estimators": kwargs["n_estimators"],
-        "criterion": kwargs["criterion"],
-        "max_depth": kwargs["max_depth"],
-        "max_features": kwargs["max_features"],
-        "min_samples_leaf": kwargs["min_samples_leaf"],
+        "C": kwargs["C"],
+        "kernel": kwargs["kernel"],
+        "gamma": kwargs["gamma"],
     }
 
     # Realizar búsqueda de hiperparámetros usando GridSearchCV
     grid_search = GridSearchCV(
-        estimator=RandomForestClassifier(random_state=kwargs["seed"]),
+        estimator=SVC(random_state=kwargs["seed"]),
         param_grid=parameters,
         scoring="f1_weighted",
         verbose=0,
@@ -61,18 +55,16 @@ def train_random_forest(X, Y, **kwargs):
     grid_search.fit(X, Y)
 
     # Obtener los mejores hiperparámetros encontrados
-    best_random_forest_model = grid_search.best_estimator_
+    best_svm_model = grid_search.best_estimator_
     best_params = grid_search.best_params_
 
-    return best_random_forest_model, best_params
+    return best_svm_model, best_params
 
 
 def main():
     if len(sys.argv) != 3:
         sys.stderr.write("Arguments error. Usage:\n")
-        sys.stderr.write(
-            "\tpython train.py preprocessed-dir-path model-dir-path\n"
-        )
+        sys.stderr.write("\tpython train.py preprocessed-dir-path model-dir-path\n")
         sys.exit(1)
 
     in_path = sys.argv[1]  # Ruta de los datos preprocesados
@@ -90,42 +82,39 @@ def main():
     y_train = train_data["0"]
 
     # Cargar los parámetros del archivo params.yaml
-    params = yaml.safe_load(open("params.yaml"))["train_random_forest"]
+    params = yaml.safe_load(open("params.yaml"))["train_svm"]
     # version_label = yaml.safe_load(open("params.yaml"))["version_label"]
     mlflow_params = yaml.safe_load(open("params.yaml"))["mlflow"]
+
+    # Inicializar la clase ClassifierModel con la versión y semilla
+    # classifier_model = trainSVM(X_train, y_train, **params)
 
     # Inicializar MLFlow
     mlflow.set_experiment(mlflow_params["experiment_name"])
     mlflow.set_tracking_uri(mlflow_params["tracking_uri"])
 
     # Asignar el nombre del run usando el formato definido en params.yaml
-    run_name = mlflow_params["run_name"].format(model="Random Forest")
+    run_name = mlflow_params["run_name"].format(model="SVM")
 
-    # Entrenar y guardar el modelo RandomForest usando los parámetros de
-    # `params.yaml`
+    # Entrenar y guardar el modelo SVM usando los parámetros de `params.yaml`
     with mlflow.start_run(run_name=run_name) as run:
-        random_forest_model, best_params = train_random_forest(
-            X_train, y_train, **params
-        )
+        svm_model, best_params = trainSVM(X_train, y_train, **params)
 
         # Guardar los mejores hiperparámetros en MLflow
         mlflow.log_params(best_params)
 
         # Guardar el modelo en MLFlow
-        mlflow.sklearn.log_model(random_forest_model, "RF_model")
+        mlflow.sklearn.log_model(svm_model, "SVM_model")
 
         # Guardar el run_id para compartirlo con evaluate.py
         run_id = run.info.run_id
-        with open(
-            os.path.join(out_path, "mlflow_run_id_random_forest_model.txt"),
-            "w",
-        ) as f:
+        with open(os.path.join(out_path, "mlflow_run_id_svm_model.txt"), "w") as f:
             f.write(run_id)
 
     # Guardar el modelo en el directorio de salida
-    model_output_path = os.path.join(out_path, "random_forest_model.pkl")
-    joblib.dump(random_forest_model, model_output_path)
-    sys.stdout.write(f"Modelo Random Forest guardado en {model_output_path}\n")
+    model_output_path = os.path.join(out_path, "svm_model.pkl")
+    joblib.dump(svm_model, model_output_path)
+    sys.stderr.write(f"Modelo SVM guardado en {model_output_path}\n")
 
 
 if __name__ == "__main__":
