+++ tests/.gitignore
@@ -1,13 +1,6 @@
-#MLFlow
-mlruns/
-model/mlflow_run_id.txt
-
-#Tests
-dump/
-
-
 # Byte-compiled / optimized / DLL files
 __pycache__/
+**/__pycache__/
 *.py[cod]
 *$py.class
