+++ tests/test_modelclassifier.py
@@ -6,15 +6,13 @@ from pandas import DataFrame
 from sklearn.metrics import classification_report
 from sklearn.model_selection import train_test_split
 
-from pipeline.model.classifiermodel import ClassifierModel
+from maternalhealth.pipeline.model.classifiermodel import ClassifierModel
 
 
 class TestModelClassifier:
     ACCURACY_THRESHOLD = 0.75
     PRECISION_THRESHOLD = 0.75
-    _tag = "tester_{date}".format(
-        date=datetime.today().strftime("%B%d_%Y__%H%M%S")
-    )
+    _tag = "tester_{date}".format(date=datetime.today().strftime("%B%d_%Y__%H%M%S"))
     _model = None
     _dataframe = None
     _X = None
@@ -82,9 +80,7 @@ class TestModelClassifier:
         hasOutliers = True
         outliersBeforePreporcessing = self.getOutliersFromDataframe(dataframe)
         dfWithNoOutliers = self._model.outlierHandler(dataframe)
-        outliersAfterPreporcessing = self.getOutliersFromDataframe(
-            dfWithNoOutliers
-        )
+        outliersAfterPreporcessing = self.getOutliersFromDataframe(dfWithNoOutliers)
 
         for column in dataframe.select_dtypes(include=[np.number]).columns:
             print(
@@ -156,9 +152,7 @@ class TestModelClassifier:
             self._model = ClassifierModel(self._tag, asTesting=True)
 
         Xt = self._model.normalizeData(
-            self._model.applyTransformations(
-                self._model.outlierHandler(self._X)
-            )
+            self._model.applyTransformations(self._model.outlierHandler(self._X))
         )
 
         yt, _ = self._model.labelEncoding(self._y, True)
@@ -187,9 +181,7 @@ class TestModelClassifier:
             self._model = ClassifierModel(self._tag, asTesting=True)
 
         Xt = self._model.normalizeData(
-            self._model.applyTransformations(
-                self._model.outlierHandler(self._X)
-            )
+            self._model.applyTransformations(self._model.outlierHandler(self._X))
         )
         yt, _ = self._model.labelEncoding(self._y, True)
